salon_scheduler/requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
Flask==3.0.0
SQLAlchemy==2.0.23
email-validator==2.1.0.post1
python-dateutil==2.8.2
Jinja2==3.1.2
pytz==2023.3
aiosmtplib==2.0.2 